# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AutoGenBook API
  description: HTTP API for AutoGenBook, served by this FastAPI app and reached through
    the nginx-fronted Docker Compose stack. Every route is mounted under /api/v1;
    there is no authentication and no CORS middleware. Error responses use RFC 9457
    problem details (application/problem+json); list endpoints return a {items, total,
    limit, offset} page envelope.
  version: 0.1.0
paths:
...
  /api/v1/projects/{project_id}/sources:
    get:
      tags:
      - sources
      summary: List Sources
      operationId: sources-list_sources
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Source_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - sources
      summary: Add Source
      operationId: sources-add_source
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/sources/{source_id}:
    get:
      tags:
      - sources
      summary: Get Source
      operationId: sources-get_source
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - sources
      summary: Update Source
      operationId: sources-update_source
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - sources
      summary: Remove Source
      operationId: sources-remove_source
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Page_Source_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Source'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[Source]
...
    Source:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        fileId:
          type: string
          format: uuid
          title: Fileid
        name:
          type: string
          title: Name
        sizeBytes:
          type: integer
          title: Sizebytes
        type:
          $ref: '#/components/schemas/SourceType'
        chunksCount:
          anyOf:
          - type: integer
          - type: 'null'
          title: Chunkscount
        status:
          $ref: '#/components/schemas/SourceStatus'
        uploadDate:
          type: string
          format: date-time
          title: Uploaddate
        authors:
          anyOf:
          - type: string
          - type: 'null'
          title: Authors
        year:
          anyOf:
          - type: string
          - type: 'null'
          title: Year
        doi:
          anyOf:
          - type: string
          - type: 'null'
          title: Doi
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - id
      - fileId
      - name
      - sizeBytes
      - type
      - status
      - uploadDate
      title: Source
    SourceCreate:
      properties:
        fileId:
          type: string
          format: uuid
          title: Fileid
        type:
          anyOf:
          - $ref: '#/components/schemas/SourceType'
          - type: 'null'
        authors:
          anyOf:
          - type: string
          - type: 'null'
          title: Authors
        year:
          anyOf:
          - type: string
            pattern: ^\d{4}$
          - type: 'null'
          title: Year
        doi:
          anyOf:
          - type: string
          - type: 'null'
          title: Doi
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - fileId
      title: SourceCreate
    SourceStatus:
      type: string
      enum:
      - ready
      - indexed
      - error
      title: SourceStatus
    SourceType:
      type: string
      enum:
      - pdf
      - doc
      - ppt
      - md
      - txt
      - slides
      - arxiv
      - notes
      - bibtex
      - latex
      - url
      - book
      - dataset
      title: SourceType
    SourceUpdate:
      properties:
        type:
          anyOf:
          - $ref: '#/components/schemas/SourceType'
          - type: 'null'
        authors:
          anyOf:
          - type: string
          - type: 'null'
          title: Authors
        year:
          anyOf:
          - type: string
            pattern: ^\d{4}$
          - type: 'null'
          title: Year
        doi:
          anyOf:
          - type: string
          - type: 'null'
          title: Doi
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      title: SourceUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError